Mistakenly added unsupported image formats
Apologies for this, partially reverting #1 from 2 days ago. Adding il32 and it32 wasn't the correct fix as the image loader does not support masked RGB. It seemed to improve rendering because of issues in the parser which are now fixed.
Removing these two lines leaves the same success in parsing and rendering without attempting to load images that will fail as it's not PNG/Jpeg data it will find.